[Refactor] #100 placeId Type 변경

+++ DaOnGil/data/src/main/java/kr/tekit/lion/data/dto/response/searchplace/auto_complete_keyword/Data.kt
@@ -8,5 +8,5 @@ data class Data(
     @Json(name = "keyword")
     val keyword: String,
     @Json(name = "placeId")
-    val placeId: Int
+    val placeId: Long
 )
